feat(memory): update func
