+++ 01-features/04-manage-context-of-your-agent/memory/06-production-patterns/00-multi-region-replication/agentcore_replication/__init__.py
@@ -25,6 +25,7 @@ from .stream_consumer import (
     make_target_client,
     process_kinesis_records,
     replicate_stream_event,
+    stream_delivery_resources,
 )
 from .dual_writer import DualRegionEventWriter
 
@@ -34,6 +35,7 @@ __all__ = [
     "make_target_client",
     "process_kinesis_records",
     "replicate_stream_event",
+    "stream_delivery_resources",
     # STM via dual-write CreateEvent (extractionMode="SKIP" on target)
     "DualRegionEventWriter",
 ]
